Rename testing schema package
Matches the structure of the samples, but with an emphasis on the fact that this is for testing purposes.
