+++ redwood-compose/src/commonTest/kotlin/app/cash/redwood/compose/ChangeListenerTest.kt
@@ -29,16 +29,16 @@ import app.cash.redwood.widget.MutableListChildren
 import assertk.assertThat
 import assertk.assertions.containsExactly
 import assertk.assertions.isEmpty
-import example.redwood.compose.Button
-import example.redwood.compose.Row
-import example.redwood.compose.ScopedRow
-import example.redwood.compose.TestSchemaProtocolBridge
-import example.redwood.compose.TestScope
-import example.redwood.compose.Text
-import example.redwood.widget.TestSchemaProtocolNodeFactory
-import example.redwood.widget.TestSchemaTestingWidgetFactory
-import example.redwood.widget.TestSchemaWidgetFactories
-import example.redwood.widget.TestSchemaWidgetFactory
+import com.example.redwood.testing.compose.Button
+import com.example.redwood.testing.compose.Row
+import com.example.redwood.testing.compose.ScopedRow
+import com.example.redwood.testing.compose.TestSchemaProtocolBridge
+import com.example.redwood.testing.compose.TestScope
+import com.example.redwood.testing.compose.Text
+import com.example.redwood.testing.widget.TestSchemaProtocolNodeFactory
+import com.example.redwood.testing.widget.TestSchemaTestingWidgetFactory
+import com.example.redwood.testing.widget.TestSchemaWidgetFactories
+import com.example.redwood.testing.widget.TestSchemaWidgetFactory
 import kotlin.test.Test
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.test.runTest
