+++ redwood-testing/src/commonTest/kotlin/app/cash/redwood/testing/ViewTreesTest.kt
@@ -34,13 +34,13 @@ import app.cash.redwood.protocol.widget.ProtocolBridge
 import app.cash.redwood.widget.MutableListChildren
 import assertk.assertThat
 import assertk.assertions.isEqualTo
-import example.redwood.compose.Row
-import example.redwood.compose.TestSchemaProtocolBridge
-import example.redwood.compose.Text
-import example.redwood.widget.TestSchemaProtocolNodeFactory
-import example.redwood.widget.TestSchemaTester
-import example.redwood.widget.TestSchemaTestingWidgetFactory
-import example.redwood.widget.TestSchemaWidgetFactories
+import com.example.redwood.testing.compose.Row
+import com.example.redwood.testing.compose.TestSchemaProtocolBridge
+import com.example.redwood.testing.compose.Text
+import com.example.redwood.testing.widget.TestSchemaProtocolNodeFactory
+import com.example.redwood.testing.widget.TestSchemaTester
+import com.example.redwood.testing.widget.TestSchemaTestingWidgetFactory
+import com.example.redwood.testing.widget.TestSchemaWidgetFactories
 import kotlin.test.Test
 import kotlinx.coroutines.plus
 import kotlinx.coroutines.test.runTest
