+++ redwood-tooling-codegen/src/test/kotlin/app/cash/redwood/tooling/codegen/ComposeProtocolGenerationTest.kt
@@ -22,7 +22,7 @@ import app.cash.redwood.tooling.schema.ProtocolSchemaSet
 import assertk.all
 import assertk.assertThat
 import assertk.assertions.contains
-import example.redwood.TestSchema
+import com.example.redwood.testing.TestSchema
 import org.junit.Test
 
 class ComposeProtocolGenerationTest {
