+++ redwood-tooling-codegen/src/test/kotlin/app/cash/redwood/tooling/codegen/WidgetProtocolGenerationTest.kt
@@ -23,7 +23,7 @@ import assertk.all
 import assertk.assertThat
 import assertk.assertions.contains
 import assertk.assertions.containsMatch
-import example.redwood.TestSchema
+import com.example.redwood.testing.TestSchema
 import kotlin.text.RegexOption.MULTILINE
 import org.junit.Test
