+++ redwood-tooling-schema/src/test/kotlin/app/cash/redwood/tooling/schema/SchemaParserTest.kt
@@ -41,9 +41,9 @@ import assertk.assertions.isInstanceOf
 import assertk.assertions.isNotNull
 import assertk.assertions.isTrue
 import assertk.assertions.message
+import com.example.redwood.testing.TestSchema
 import com.google.testing.junit.testparameterinjector.TestParameter
 import com.google.testing.junit.testparameterinjector.TestParameterInjector
-import example.redwood.TestSchema
 import java.io.File
 import kotlin.DeprecationLevel.HIDDEN
 import kotlin.reflect.KClass
@@ -756,7 +756,7 @@ class SchemaParserTest(
     assertFailure { parser.parse(SchemaDependencyHasDependency::class) }
       .isInstanceOf<IllegalArgumentException>()
       .hasMessage(
-        "Schema dependency example.redwood.TestSchema also has its own dependencies. " +
+        "Schema dependency com.example.redwood.testing.TestSchema also has its own dependencies. " +
           "For now, only a single level of dependencies is supported.",
       )
   }
