+++ test-app/schema/modifiers/build.gradle
@@ -27,5 +27,5 @@ kotlin {
 
 redwoodSchema {
   source = projects.testApp.schema
-  type = 'example.redwood.TestSchema'
+  type = 'com.example.redwood.testing.TestSchema'
 }
