+++ test-app/schema/testing/build.gradle
@@ -39,5 +39,5 @@ kotlin {
 
 redwoodSchema {
   source = projects.testApp.schema
-  type = 'example.redwood.TestSchema'
+  type = 'com.example.redwood.testing.TestSchema'
 }
