+++ test-app/schema/widget/build.gradle
@@ -36,5 +36,5 @@ kotlin {
 
 redwoodSchema {
   source = projects.testApp.schema
-  type = 'example.redwood.TestSchema'
+  type = 'com.example.redwood.testing.TestSchema'
 }
